address pr review: move emit_* helpers to events.rs; clarify channel comments
